+++ src/main/java/com/cloudDomus/cloudDommus/Manager/ManagerController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @RestController
 @RequestMapping("/api/managers")
@@ -33,14 +34,24 @@ public class ManagerController {
     @GetMapping("/manager/{id}")
     public Manager getManagerByID(@PathVariable Long id) {
 
-        return repository.findById(id).orElseThrow(() -> new ManagerNotFoundException(id));
+        return repository.getById(id);
     }
 
     @ApiOperation(value = "Delete a manager by its Id", response = List.class)
     @DeleteMapping("/Manager/{id}")
-    public void deleteManager(@PathVariable Long id) {
+    @Transactional(rollbackFor = ManagerNotFoundException.class)
+    public Manager deleteManager(@PathVariable Long id) {
+        Manager deleted = repository.getById(id);
+        if(deleted == null){
+            throw new ManagerNotFoundException(id);
+        }
         repository.deleteById(id);
+        deleted.setId(null);
+        return deleted;
     }
 
+    public void setRepository(ManagerRepository repository) {
+        this.repository = repository;
+    }
 
 }
