+++ src/main/java/com/cloudDomus/cloudDommus/Service/ServiceController.java
@@ -3,6 +3,7 @@ package com.cloudDomus.cloudDommus.Service;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,7 +16,6 @@ public class ServiceController {
     @Autowired
     ServiceRepository repository;
 
-    // Aggregate root
     @ApiOperation(value = "View a list of available Service", response = List.class)
     @GetMapping("/service")
     public List<Service> all() {
@@ -28,18 +28,26 @@ public class ServiceController {
         return repository.save(newService);
     }
 
-    // Single item
     @ApiOperation(value = "Get service by Id", response = List.class)
     @GetMapping("/service/{id}")
     public Service getServiceByID(@PathVariable Long id) {
-
-        return repository.findById(id).orElseThrow(() -> new ServiceNotFoundException(id));
+        return repository.getById(id);
     }
 
     @ApiOperation(value = "Delete a service by its Id", response = List.class)
     @DeleteMapping("/Service/{id}")
-    public void deleteService(@PathVariable Long id) {
+    @Transactional(rollbackFor = ServiceNotFoundException.class)
+    public Service deleteService(@PathVariable Long id) throws ServiceNotFoundException{
+        Service deleted = repository.getById(id);
+        if(deleted == null){
+            throw new ServiceNotFoundException(id);
+        }
         repository.deleteById(id);
+        deleted.setId(null);
+        return deleted;
     }
 
+    public void setRepository(ServiceRepository repository) {
+        this.repository = repository;
+    }
 }
